+++ CHANGELOG.md
@@ -20,7 +20,10 @@
 - Bugfix: The cluster domain used by the DNS resolver is retrieved from the traffic-manager instead of being
   hard-coded to "cluster.local".
 
+- Bugfix: "Telepresence uninstall --everything" now also uninstalls agents installed via mutating webhook
+
 - Bugfix: Downloading large files during an intercept will no longer cause timeouts and hanging traffic-agent.
+
 
 ### 2.4.2 (September 1, 2021)